+++ src/ingest_validation_tools/upload.py
@@ -34,6 +34,7 @@ from ingest_validation_tools.schema_loader import (
     get_table_schema,
 )
 from ingest_validation_tools.validation_utils import (
+    TSVError,
     cedar_validation_call,
     find_empty_tsv_columns,
     get_data_dir_errors,
@@ -75,6 +76,7 @@ class Upload:
     ):
         del kwargs
         self.directory_path = directory_path
+        self.specified_paths = tsv_paths
         # typical dataset_ignore_globs = [uuid, "extras", "*metadata.tsv", "validation_report.txt"]
         self.dataset_ignore_globs = dataset_ignore_globs
         # typical upload_ignore_globs = "*"
@@ -146,12 +148,11 @@ class Upload:
         self.get_info_called = True
         return self.info
 
-    def get_errors(self, **kwargs) -> ErrorDict:
+    def get_errors(self) -> ErrorDict:
         """
         When converted using ErrorDict.as_dict(), keys are
         present only if there is actually an error to report.
         """
-        # TODO: remove deprecated kwargs param
         # Return if PreflightErrors found
         if self.errors:
             return self.errors
@@ -185,7 +186,7 @@ class Upload:
             k: unsorted_tsv_paths[k] for k in sorted(unsorted_tsv_paths.keys())
         }
         if not self.dataset_metadata:
-            self.errors.preflight.value = "There are no metadata TSVs."
+            raise PreflightError("There are no metadata TSVs.")
 
     def get_app_context(self, submitted_app_context: dict):
         """
@@ -986,26 +987,49 @@ class Upload:
     def _check_for_misnamed_metadata_tsvs(self):
         """
         Raise PreflightError if any TSVs are located in the top-level directory that
-        do not end with "-metadata.tsv" to avoid mistakenly validating a multi-assay upload
+        do not end with "-metadata.tsv" / are not an antibodies/contributors path
+        in order to avoid mistakenly validating a multi-assay upload
         as a single-assay upload (e.g. in the case of misnamed metadata files).
         """
-        other_tsvs = [
+        # tsv_paths specified for upload, do not check parent directory
+        if self.specified_paths:
+            return
+        # identify expected antibodies/contributors paths
+        other_paths = []
+        for other_type in ["antibodies_path", "contributors_path"]:
+            referenced_paths = self.__get_referenced_paths(other_type)
+            for path_value in referenced_paths:
+                other_paths.append(self.directory_path / path_value)
+
+        extra_tsvs = [
             p
             for p in self.directory_path.glob("*.tsv")
-            if not p.name.endswith(TSV_SUFFIX) and p not in self.dataset_metadata
+            if not p.name.endswith(TSV_SUFFIX) and p not in [*self.dataset_metadata, *other_paths]
         ]
 
+        # compile extra TSVs and any errors during file read
         extra_metadata_tsvs = []
-        for path in other_tsvs:
-            rows = read_rows(path, self.encoding)
+        extra_errors = []
+        for path in extra_tsvs:
+            try:
+                rows = read_rows(path, self.encoding)
+            except TSVError as e:
+                extra_errors.append(str(e))
+                continue
+            # we only want to highlight misnamed metadata files here;
+            # reference checking should catch any other extra TSVs
             if "dataset_type" in rows[0]:
                 extra_metadata_tsvs.append(path)
 
-        if extra_metadata_tsvs:
-            names = ", ".join(p.name for p in sorted(extra_metadata_tsvs))
-            raise PreflightError(
-                f"Metadata TSV file(s) found that do not end in '-{TSV_SUFFIX}': {names}."
-            )
+        if extra_metadata_tsvs or extra_errors:
+            message = []
+            if names := ", ".join(p.name for p in sorted(extra_metadata_tsvs)):
+                message.append(
+                    f"Metadata TSV file(s) found that do not end in '-{TSV_SUFFIX}': {names}."
+                )
+            if extra_errors:
+                message.append(", ".join(extra_errors))
+            raise PreflightError(" ".join(message))
 
     def _check_multi_assay(self):
         # first make sure there are no top-level metadata TSVs that are not accounted for