+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Generate DOI objects
 - Patch for TSVs not being caught
 - Update Illumina Spatial (v0) directory schema
+- Update CosMx Transcriptomics directory schema
 
 ## v1.1.4