+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Patch for TSVs not being caught
 - Update Illumina Spatial (v0) directory schema
 - Update CosMx Transcriptomics directory schema
+- Update RNAseq (with probes) directory schema
 
 ## v1.1.4